requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
faiss-cpu==1.8.0
Flask==3.0.3
frozenlist==1.4.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.49
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.23.1
orjson==3.10.1
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pypdf==4.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.29
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
faiss-cpu==1.8.0
Flask==3.0.3
frozenlist==1.4.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.49
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.23.1
orjson==3.10.1
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pypdf==4.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.29
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
